// src/common/enums.ts
export enum WorkType {
  SingleOrder = 0,
  GlobalOrder = 1,
  SingleParallel = 2,
  GlobalParallel = 3
}

export enum AccountType {
  Normal = 0,
  Co = 1
}

export enum Server {
  Australia = 0,
  England = 1,
  Arabic = 2,
  Bulgaria = 3,
  International = 4,
  Spain = 5,
  Canada = 6,
  Germany = 7,
  Norway = 8,
  Poland = 9,
  Russia = 10,
  Romain = 11,
  USA = 12,
  FranceOuranos = 13,
  FranceGaia = 14,
  Sweden = 15,
}

export enum HorseSex { Male = 0, Female = 1 }

export enum ClientType { New = 0, Old = 1 }

export enum Result { None = 0, Success = 1, Error = 2, Empty = 3 }

export enum ProductType {
  None = 0,
  Hay = 1,
  Oat = 2,
  Wheat = 3,
  Shit = 4,
  Leather = 5,
  Apples = 6,
  Carrot = 7,
  Wood = 8,
  Steel = 9,
  Sand = 10,
  Straw = 11,
  Flax = 12,
  OR = 13
}

export enum Localization { ru = 0, en = 1 }